protect Vistle TCP connections with HMAC
VISTLE_KEY environment variable should be set to hexadecimal string of
session key. Session key will be used for creating and verifying HMAC of
initial message on every connection.

If VISTLE_KEY is not set, it will be set to random data, thus enabling a
single-cluster session.

Botan is required for cryptographic routines.