create VtkmModule class to make it easier to port Vistle module to the GPU (#253)
Modules that use VTK-m to run algorithms on different devices, especially on the GPU, can now inherit from the VtkmModule class which handles reading in the data on the input ports, converting between the VTK-m and Vistle data format, as well as writing the filter results to the output ports.
Child classes must override the setUpFilter method which sets up the desired VTK-m filter. They are, optionally, free to make changes to the methods preparing the filter's input dataset (prepareInputGrid and prepareInputField) and/or the methods preparing the filter output that is to be added to the output ports (prepareOutputGrid and prepareOutputField).

Co-authored-by: Martin Aumüller <aumuell@reserv.at>