add script build_check:
This script is meant to be a tool to check the building process in an encapsulated environment like a server for example. It first checks the build status of the github action of the specified github repository. In case of valid checks it will proceed to fetch the latest changes and starts the configuring process of cmake followed by starting the compile process with make. If any of the processes fail to run successfully the script will stop and print out the error log in /tmp/build_check.log

How to use:

Put this script or a symbolic link of it in your preferred
location and adjust the setup in the script itself. Make it executable
and run it

chmod +x build_check.sh
./daily_check.sh